Link: faithful RESOURCE_ADV dispatch with request/response routing
Port the upstream RESOURCE_ADV handling order (Link.py:1069-1102):
resource-borne REQUESTs are accepted with the link's
request_resource_concluded handler, resource-borne RESPONSEs are
matched against pending requests by request_id and accepted with
response_resource_concluded (updating response_size /
response_transfer_size / started_at and firing the initial progress
callback as upstream does), and only plain resources fall through to
the resource_strategy gate. ACCEPT_NONE now silently ignores the
advertisement after decrypt (upstream 'pass'); ACCEPT_APP invokes the
link's resource accept callback (now returning bool, with the
advertisement's link set for the duration of the call) and sends
RESOURCE_RCL when the application declines, mirroring Resource.reject;
ACCEPT_ALL accepts as before. Accepted resources also fire the link's
resource_started callback (Resource.py:225-229).

The embedded guard checks (compression/metadata refusal, zero size,
runtime max-incoming cap, flash quota) now gate every accepted
resource including request/response resources — a deliberate
divergence from upstream, which accepts those ungated.

Resource callbacks are plain function pointers, so the bound-method
callbacks upstream passes into Resource.accept become static
trampolines that recover the owning Link from the resource.

Fixes drift found while wiring the handlers: Resource::data() now
returns the assembled cleartext on the receiver (the handlers hashed/
unpacked an empty buffer before), RequestReceipt progress is stored
as float (an int truncated every fractional progress value to 0), and
response_resource_concluded snapshots matching receipts before
request_timed_out erases them from the set it was iterating.

+++ src/LinkData.h
@@ -145,7 +145,9 @@ namespace RNS {
 		size_t _response_size = 0;
 		Type::RequestReceipt::status _status = Type::RequestReceipt::SENT;
 		double _sent_at = 0.0;
-		int _progress = 0;
+		// 0.0..1.0, mirrors upstream RequestReceipt.progress (set from
+		// resource.get_progress(), a float — an int here truncated to 0).
+		float _progress = 0.0f;
 		double _concluded_at = 0.0;
 		double _response_concluded_at = 0.0;
 		double _timeout = 0.0;
